@@ -59,6 +59,7 @@ import * as Tasty from '../../src/tasty/Tasty';
 let elasticDB: ElasticDB;
 let server;
 
+let defaultUserAccessToken: string = '';
 let exportTemplateID: number = -1;
 let persistentExportAccessToken: string = '';
 
@@ -181,11 +182,27 @@ beforeAll(async (done) =>
     fail(e);
   }
 
-  request(server)
+  await request(server)
+    .post('/midway/v1/auth/login')
+    .send({
+      email: 'admin@terraindata.com',
+      password: 'CnAATPys6tEB*ypTvqRRP5@2fUzTuY!C^LZP#tBQcJiC*5',
+    })
+    .then((response) =>
+    {
+      const respData = JSON.parse(response.text);
+      defaultUserAccessToken = respData.accessToken;
+    })
+    .catch((error) =>
+    {
+      winston.warn('Error while creating access token for default user: ' + String(error));
+    });
+
+  await request(server)
     .post('/midway/v1/users/')
     .send({
       id: 1,
-      accessToken: 'ImAnAdmin',
+      accessToken: defaultUserAccessToken,
       body: {
         email: 'test@terraindata.com',
         name: 'Test Person',
@@ -195,9 +212,13 @@ beforeAll(async (done) =>
         timezone: 'UTC',
       },
     })
-    .end(() =>
+    .then((response) =>
     {
       done();
+    })
+    .catch((error) =>
+    {
+      winston.warn('Error while creating test user: ' + String(error));
     });
 
   try
@@ -234,7 +255,7 @@ describe('Status tests', () =>
       .get('/midway/v1/status/stats')
       .query({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
       })
       .expect(200)
       .then((response) =>
@@ -373,7 +394,7 @@ describe('Item route tests', () =>
       .get('/midway/v1/items/')
       .query({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
       })
       .expect(200)
       .then((response) =>
@@ -413,7 +434,7 @@ describe('Item route tests', () =>
       .post('/midway/v1/items/')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           name: 'Test Item',
           status: 'LIVE',
@@ -443,7 +464,7 @@ describe('Item route tests', () =>
       .get('/midway/v1/items/1')
       .query({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
       })
       .expect(200)
       .then((response) =>
@@ -473,7 +494,7 @@ describe('Item route tests', () =>
       .post('/midway/v1/items/2')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: insertObject,
       })
       .expect(200)
@@ -496,7 +517,7 @@ describe('Item route tests', () =>
       .post('/midway/v1/items/314159265359')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           id: 314159265359,
           name: 'Test Item',
@@ -569,7 +590,7 @@ describe('Schema route tests', () =>
       .get('/midway/v1/schema/')
       .query({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
       })
       .expect(200)
       .then((response) =>
@@ -591,7 +612,7 @@ describe('Query route tests', () =>
       .post('/midway/v1/query/')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           database: 1,
           type: 'search',
@@ -626,7 +647,7 @@ describe('Query route tests', () =>
       .post('/midway/v1/query/')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           database: 1,
           type: 'wrongtype',
@@ -677,7 +698,7 @@ describe('Query route tests', () =>
         .post('/midway/v1/query/')
         .send({
           id: 1,
-          accessToken: 'ImAnAdmin',
+          accessToken: defaultUserAccessToken,
           body: {
             database: 1,
             type: 'putTemplate',
@@ -698,7 +719,7 @@ describe('Query route tests', () =>
         .post('/midway/v1/query/')
         .send({
           id: 1,
-          accessToken: 'ImAnAdmin',
+          accessToken: defaultUserAccessToken,
           body: {
             database: 1,
             type: 'getTemplate',
@@ -729,7 +750,7 @@ describe('Query route tests', () =>
         .post('/midway/v1/query/')
         .send({
           id: 1,
-          accessToken: 'ImAnAdmin',
+          accessToken: defaultUserAccessToken,
           body: {
             database: 1,
             type: 'deleteTemplate',
@@ -751,7 +772,7 @@ describe('Query route tests', () =>
         .post('/midway/v1/query/')
         .send({
           id: 1,
-          accessToken: 'ImAnAdmin',
+          accessToken: defaultUserAccessToken,
           body: {
             database: 1,
             type: 'getTemplate',
@@ -787,7 +808,7 @@ describe('Query route tests', () =>
       .post('/midway/v1/query/')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           database: 1,
           type: 'search',
@@ -862,7 +883,7 @@ describe('Query route tests', () =>
       .post('/midway/v1/query/')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           database: 1,
           type: 'search',
@@ -945,7 +966,7 @@ describe('File import route tests', () =>
   {
     await request(server)
       .post('/midway/v1/import/')
-      .field('accessToken', 'ImAnAdmin')
+      .field('accessToken', defaultUserAccessToken)
       .field('columnTypes', JSON.stringify({
         pkey: { type: 'long' },
         col1: { type: 'text' },
@@ -1008,7 +1029,7 @@ describe('File import route tests', () =>
   {
     await request(server)
       .post('/midway/v1/import/')
-      .field('accessToken', 'ImAnAdmin')
+      .field('accessToken', defaultUserAccessToken)
       .field('columnTypes', JSON.stringify({
         pkey: { type: 'long' },
         col1: { type: 'text' },
@@ -1089,7 +1110,7 @@ describe('File import route tests', () =>
   {
     await request(server)
       .post('/midway/v1/import/')
-      .field('accessToken', 'ImAnAdmin')
+      .field('accessToken', defaultUserAccessToken)
       .field('columnTypes', JSON.stringify({
         pkey: { type: 'long' },
         col1: { type: 'text' },
@@ -1134,7 +1155,7 @@ describe('File io templates route tests', () =>
       .post('/midway/v1/import/templates/create')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           name: 'mysql_import_template',
           dbid: 1,
@@ -1254,7 +1275,7 @@ describe('File io templates route tests', () =>
       .post('/midway/v1/export/templates/create')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           name: 'my_template',
           dbid: 1,
@@ -1310,7 +1331,7 @@ describe('File io templates route tests', () =>
       .get('/midway/v1/export/templates/')
       .query({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
       })
       .expect(200)
       .then((response) =>
@@ -1349,7 +1370,7 @@ describe('File io templates route tests', () =>
       .post('/midway/v1/export/templates/')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           dbid: 1,
           dbname: 'badname',
@@ -1561,7 +1582,7 @@ describe('Credentials tests', () =>
       .post('/midway/v1/credentials')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           createdBy: 1,
           meta: '"{\"host\":\"10.1.1.103\", \"port\":22, \"username\":\"testuser\", \"password\":\"Terrain123!\"}"',
@@ -1596,7 +1617,7 @@ describe('Credentials tests', () =>
       .get('/midway/v1/credentials')
       .query({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
       })
       .expect(200)
       .then((response) =>
@@ -1637,7 +1658,7 @@ describe('Scheduler tests', () =>
       .post('/midway/v1/scheduler/create')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           jobType: 'export',
           schedule: '* * * * *', // next run on some leap year date
@@ -1688,7 +1709,7 @@ describe('Scheduler tests', () =>
       .post('/midway/v1/scheduler/run/' + schedulerExportId.toString())
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
         },
       })
@@ -1712,7 +1733,7 @@ describe('Scheduler tests', () =>
       .post('/midway/v1/scheduler/create')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           jobTypeInvalidParam: 'export',
           schedule: '* * * * *', // next run on some leap year date
@@ -1757,7 +1778,7 @@ describe('Analytics events route tests', () =>
       .get('/midway/v1/events/agg')
       .query({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         database: 1,
         agg: 'distinct',
       })
@@ -1784,7 +1805,7 @@ describe('Analytics route tests', () =>
       .get('/midway/v1/events/agg')
       .query({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         database: 1,
         start: new Date(2018, 2, 16, 7, 24, 4),
         end: new Date(2018, 2, 16, 7, 36, 4),
@@ -1811,7 +1832,7 @@ describe('Analytics route tests', () =>
       .get('/midway/v1/events/agg')
       .query({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         database: 1,
         start: new Date(2018, 2, 16, 7, 24, 4),
         end: new Date(2018, 2, 16, 7, 36, 4),
@@ -1839,7 +1860,7 @@ describe('Analytics route tests', () =>
       .get('/midway/v1/events/agg')
       .query({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         database: 1,
         start: new Date(2018, 3, 3, 7, 24, 4),
         end: new Date(2018, 3, 3, 10, 24, 4),
@@ -1867,7 +1888,7 @@ describe('Analytics route tests', () =>
       .post('/midway/v1/events/metrics')
       .send({
         id: 1,
-        accessToken: 'ImAnAdmin',
+        accessToken: defaultUserAccessToken,
         body: {
           database: 1,
           label: 'Clicks',
