@@ -85,9 +85,11 @@ class AuthRedux extends TerrainRedux<AuthActionTypes, AuthState>
 
       logout: (state, action) =>
       {
+        const storedAccessToken = localStorage['accessToken'];
+        const storedId = localStorage['id'];
         localStorage.removeItem('accessToken');
         localStorage.removeItem('id');
-        Ajax.logout(_.noop);
+        Ajax.logout(storedAccessToken, storedId);
 
         return state.set('accessToken', null).set('id', null);
       },
