@@ -576,6 +576,59 @@ export class CardsToPath
     return filterLines;
   }
 
+  private static extractNestedQueryToFilter(queries: ESValueInfo, parser: ESCardParser, flip: boolean = false)
+  {
+    const filters = [];
+    queries.forEachElement((query: ESValueInfo) =>
+    {
+      if (query.objectChildren.nested)
+      {
+        const nestedQuery = query.objectChildren.nested.propertyValue;
+        const boolQuery = parser.searchCard({ 'query:query': { 'bool:elasticFilter': true } }, nestedQuery);
+        if (boolQuery !== null)
+        {
+          // create filter group
+          // let newFilterGroup = _FilterGroup();
+          boolQuery.card.otherFilters.forEach((filter) =>
+          {
+            // TODO convert boolQuery from must to filter ?
+            let line = this.TerrainFilterBlockToFilterLine(filter);
+            if (line)
+            {
+              if (flip)
+              {
+                // we only handle `exits`, `equal`, and `isin`
+                let keep = true;
+                switch (line.comparison)
+                {
+                  case 'exists':
+                    line = line.set('comparison', 'notexists');
+                    break;
+                  case 'equal':
+                    line = line.set('comparison', 'notequal');
+                    break;
+                  case 'isin':
+                    line = line.set('comparison', 'isnotin');
+                    break;
+                  default:
+                    keep = false;
+                }
+                if (keep === true)
+                {
+                  filters.push(line);
+                }
+              } else
+              {
+                filters.push(line);
+              }
+            }
+          });
+        }
+      }
+    });
+    return filters;
+  }
+
   private static processNestedQueryFilterGroup(parentFilterGroup: FilterGroup, parser: ESCardParser, parentBool: ESValueInfo, sectionType: 'hard' | 'soft')
   {
     // let search whether we have an inner bool or not
@@ -589,34 +642,25 @@ export class CardsToPath
       from = parentBool.objectChildren.should;
     }
 
-    const newLines = [];
+    let newLines = [];
     if (from)
     {
       const queries = from.propertyValue;
       if (queries.jsonType === ESJSONType.array)
       {
-        queries.forEachElement((query: ESValueInfo) =>
-        {
-          if (query.objectChildren.nested)
-          {
-            const nestedQuery = query.objectChildren.nested.propertyValue;
-            const boolQuery = parser.searchCard({ 'query:query': { 'bool:elasticFilter': true } }, nestedQuery);
-            if (boolQuery !== null)
-            {
-              // create filter group
-              // let newFilterGroup = _FilterGroup();
-              boolQuery.card.otherFilters.forEach((filter) =>
-              {
-                // TODO convert boolQuery from must to filter ?
-                const line = this.TerrainFilterBlockToFilterLine(filter);
-                if (line)
-                {
-                  newLines.push(line);
-                }
-              });
-            }
-          }
-        });
+        const r = this.extractNestedQueryToFilter(queries, parser);
+        newLines = newLines.concat(r);
+      }
+    }
+
+    // negative filter lines
+    if (parentBool.objectChildren.must_not)
+    {
+      const queries = parentBool.objectChildren.must_not.propertyValue;
+      if (queries.jsonType === ESJSONType.array)
+      {
+        const r = this.extractNestedQueryToFilter(queries, parser, true);
+        newLines = newLines.concat(r);
       }
     }
     return List(newLines);
